requirements.txt
selenium==4.18.1
pandas==2.2.1
openpyxl==3.1.2
google-cloud-aiplatform==1.42.1
python-dotenv==1.0.1